implemented BookTests
