+++ test/GetOneArticleTests.java
@@ -9,31 +9,8 @@ import static play.test.Helpers.*;
 
 public class GetOneArticleTests extends WithApplication {
 
-    @Test
-    public void getOneArticleTest(){
-        Http.RequestBuilder request = new Http.RequestBuilder().method("GET")
-                .uri("/api/articles/3");
-        Result result = route(request);
 
-        assertEquals(OK,result.status());
-    }
 
-    @Test
-    public void getOneArticleWithNonExistingIDTest(){
-        Http.RequestBuilder request = new Http.RequestBuilder().method("GET")
-                .uri("/api/articles/0");
-        Result result = route(request);
 
-        assertEquals(BAD_REQUEST,result.status());
-    }
-
-    @Test
-    public void getOneArticleWithWrongURI(){
-        Http.RequestBuilder request = new Http.RequestBuilder().method("GET")
-                .uri("/api/articl/3");
-        Result result = route(request);
-
-        assertEquals(NOT_FOUND,result.status());
-    }
 
 }
